# openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "Payment-Signature",
        "description": "x402 v2 payment payload (base64)"
      }
    }
  },
  "info": {
    "title": "x402 Trust Layer",
    "version": "5.2.1",
    "description": "68 paid x402 trust infrastructure APIs: guard, semantic escrow, mandate diff, certified seller network, KYM, mandates, compliance, disputes, and orchestration.",
...
  },
...
  "servers": [
    {
      "url": "https://x402trustlayer.xyz"
    }
  ],
  "paths": {
...
    "/api/attestation/issue": {
      "post": {
        "operationId": "post_attestation_issue",
        "summary": "Issue signed preflight attestation",
        "description": "Issue signed preflight attestation \u2014 $0.04 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "killer",
          "attestation",
          "trust",
          "security"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "attestation": {
                    "attestationId": "att_example",
                    "allowed": true
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.040000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agentId": {
                    "type": "string"
                  },
                  "walletAddress": {
                    "type": "string"
                  },
                  "targetUrl": {
                    "type": "string"
                  },
                  "estimatedCostUsdc": {
                    "type": "number"
                  },
                  "policy": {
                    "type": "object",
                    "properties": {
                      "dailyCapUsdc": {
                        "type": "number"
                      },
                      "perCallCapUsdc": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "dailyCapUsdc",
                      "perCallCapUsdc"
                    ]
                  }
                },
                "required": [
                  "agentId",
                  "walletAddress",
                  "targetUrl",
                  "estimatedCostUsdc",
                  "policy"
                ]
              },
              "example": {
                "agentId": "dexter-verifier-probe",
                "walletAddress": "9c7tE587KpGYBjiNQrjw3nGvxQHhSYKU4Ba6WRgQsHkt",
                "targetUrl": "https://api.myceliasignal.com/oracle/price/eth/usd",
                "estimatedCostUsdc": 0.03,
                "policy": {
                  "dailyCapUsdc": 10,
                  "perCallCapUsdc": 0.5
                }
              }
            }
          }
        }
      }
    },
    "/api/attestation/verify": {
      "post": {
        "operationId": "post_attestation_verify",
        "summary": "Verify attestation signature",
        "description": "Verify attestation signature \u2014 $0.02 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "killer",
          "attestation",
          "verify"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "attestation": {
                    "attestationId": "att_example",
                    "allowed": true
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attestationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "attestationId"
                ]
              },
              "example": {
                "attestationId": "att_verifier_probe_example"
              }
            }
          }
        }
      }
    },
    "/api/attestation/registry": {
      "get": {
        "operationId": "get_attestation_registry",
        "summary": "Query attestation registry",
        "description": "Query attestation registry \u2014 $0.02 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "killer",
          "attestation",
          "registry"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registry records after payment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "count": 0,
                  "records": [],
                  "policy": "Attestation registry entries"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "minGrade",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "agentId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": false
          }
        ]
      }
    },
...
    "/api/merchant-trust/score": {
      "post": {
        "operationId": "post_merchant_trust_score",
        "summary": "Know-Your-Merchant trust + wash-trading score before payment",
        "description": "Know-Your-Merchant trust + wash-trading score before payment \u2014 $0.06 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "tier1",
          "trust",
          "kym",
          "wash-trade",
          "preflight"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.060000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "host": {
                    "type": "string"
                  },
                  "targetUrl": {
                    "type": "string"
                  },
                  "observedTxns": {
                    "type": "number"
                  },
                  "observedVolumeUsdc": {
                    "type": "number"
                  },
                  "washTradePct": {
                    "type": "number"
                  },
                  "verifiedResources": {
                    "type": "number"
                  },
                  "totalResources": {
                    "type": "number"
                  },
                  "p50LatencyMs": {
                    "type": "number"
                  },
                  "probe": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "host",
                  "targetUrl",
                  "observedTxns",
                  "observedVolumeUsdc",
                  "washTradePct",
                  "verifiedResources",
                  "totalResources",
                  "p50LatencyMs",
                  "probe"
                ]
              },
              "example": {
                "host": "api.myceliasignal.com",
                "targetUrl": "https://api.myceliasignal.com/oracle/price/eth/usd",
                "observedTxns": 54,
                "observedVolumeUsdc": 1.76,
                "washTradePct": 17,
                "verifiedResources": 44,
                "totalResources": 91,
                "p50LatencyMs": 1200,
                "probe": false
              }
            }
          }
        }
      }
    },
    "/api/mandate/compile": {
      "post": {
        "operationId": "post_mandate_compile",
        "summary": "Compile a signed, scoped AP2-style payment mandate from intent",
        "description": "Compile a signed, scoped AP2-style payment mandate from intent \u2014 $0.08 USDC via x402 (Dexter facilitator https://x402.dexter.cash).",
        "tags": [
          "tier1",
          "mandate",
          "ap2",
          "intent",
          "governance"
        ],
        "security": [
          {
            "x402": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response after x402 settlement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Successful response after x402 settlement",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "allowed": {
                      "type": "boolean"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "checks_passed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "allowed": true,
                  "summary": "Paid response after x402 settlement"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required (x402 v2). Unpaid requests must receive HTTP 402 with Payment-Required header and non-empty accepts.",
            "headers": {
              "Payment-Required": {
                "description": "Base64-encoded x402 payment requirements (v2)",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.080000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "principal": {
                    "type": "string"
                  },
                  "agentId": {
                    "type": "string"
                  },
                  "intent": {
                    "type": "string"
                  },
                  "maxPerTxUsdc": {
                    "type": "number"
                  },
                  "dailyCapUsdc": {
                    "type": "number"
                  },
                  "allowedMerchants": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "allowedCategories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "allowedRails": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "ttlMinutes": {
                    "type": "number"
                  }
                },
                "required": [
                  "principal",
                  "agentId",
                  "intent",
                  "maxPerTxUsdc",
                  "dailyCapUsdc",
                  "allowedMerchants",
                  "allowedCategories",
                  "allowedRails",
                  "ttlMinutes"
                ]
              },
              "example": {
                "principal": "cardholder:dexter-verifier",
                "agentId": "dexter-verifier-probe",
                "intent": "Buy ETH/USD oracle data for a trading bot, under $1 per call, daily $10 cap",
                "maxPerTxUsdc": 0.5,
                "dailyCapUsdc": 10,
                "allowedMerchants": [
                  "myceliasignal.com",
                  "dexter.cash"
                ],
                "allowedCategories": [
                  "market-data",
                  "oracle"
                ],
                "allowedRails": [
                  "base-x402",
                  "solana-x402",
                  "visa-cli"
                ],
                "ttlMinutes": 1440
              }
            }
          }
        }
      }
    },
...
  }
}